Refactor YelpApi code and change content_main
I removed the YelpComboWrapper class since it was unnecessary and I also changed the way the YelpHtmlScrapper worked by making it work on the same thread as the YelpAPIWrapper. Also now there is only one button in order to generate and display the event/activities instead of two separate buttons.